backend/requirements.txt
# Pinned loosely for Python 3.14: pydantic-core needs a version that ships cp314 wheels
# (avoids local Rust/MSVC builds).
fastapi>=0.115.6,<0.116
uvicorn[standard]>=0.32.1,<0.33
sqlalchemy>=2.0.36,<2.1
pydantic>=2.12.0,<3
pydantic-settings>=2.7.0,<3
httpx>=0.28.1,<0.29
feedparser>=6.0.11,<7
pyyaml>=6.0.2,<7
python-dotenv>=1.0.1,<2
openai>=1.57.4,<2
